fix "sign_integer" function bug and calibrate current sensor offset

+++ User_acmsimcv5/utility.c
@@ -4,14 +4,12 @@ extern REAL sig_a3;
 
 /* 功能函数 */
 // 符号函数
-//* utility的函数定义要小心，由于仿真和emy-c的ACMSim.h的内容不同，可能会导致编译错误
-
+ //* utility的函数定义要小心，由于仿真和emy-c的ACMSim.h的内容不同，可能会导致编译错误
 REAL sign(REAL x){
    return (x > 0) - (x < 0);
 }
 
-
-REAL sign_integer(int32 x){
+int32 sign_integer(int32 x){
     return (x > 0) - (x < 0);
 }
